+++ src/components/navbar/MenuComponent.js
@@ -1,0 +1,27 @@
+import React from "react";
+import MenuItem from "@material-ui/core/MenuItem";
+import Menu from "@material-ui/core/Menu";
+
+export default function MenuComponent({
+  anchorEl,
+  menuId,
+  isMenuOpen,
+  handleMenuClose,
+}) {
+  return (
+    <>
+      <Menu
+        anchorEl={anchorEl}
+        anchorOrigin={{ vertical: "top", horizontal: "right" }}
+        id={menuId}
+        keepMounted
+        transformOrigin={{ vertical: "top", horizontal: "right" }}
+        open={isMenuOpen}
+        onClose={handleMenuClose}
+      >
+        <MenuItem onClick={handleMenuClose}>Profile</MenuItem>
+        <MenuItem onClick={handleMenuClose}>Grants</MenuItem>
+      </Menu>
+    </>
+  );
+}
